server: rename radio protocol helper module to radio api
