+++ rigflow-server/src/server/mod.rs
@@ -5,7 +5,7 @@ pub mod connection;
 pub mod discovery;
 pub mod pipeline_factory;
 pub mod radio_manager;
-pub mod radio_protocol;
+pub mod radio_api;
 pub mod radio_types;
 pub mod radio_worker;
 pub mod session;
